requirements.txt
appnope==0.1.4
astroid==3.2.2
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blowfish==0.6.1
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
construct==2.10.70
decorator==5.1.1
deepmerge==1.1.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
frida==16.2.5
frida-tools==12.4.2
fuzzysearch==0.7.3
fuzzywuzzy==0.18.0
getch==1.0
humanfriendly==10.0
idna==3.7
iGetMusic==0.1.0
ipython==8.12.3
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
Levenshtein==0.25.1
MacFSEvents==0.8.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mistune==3.0.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.4
packaging==24.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
pylint==3.2.4
pyobjc-core==10.3.1
pyobjc-framework-AppleScriptKit==10.3.1
pyobjc-framework-AppleScriptObjC==10.3.1
pyobjc-framework-ApplicationServices==10.3.1
pyobjc-framework-Cocoa==10.3.1
pyobjc-framework-CoreText==10.3.1
pyobjc-framework-Quartz==10.3.1
pyrekordbox==0.3.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.25.1
PyYAML==6.0.1
pyzmq==26.0.3
rapidfuzz==3.9.3
redis==5.0.5
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
setuptools==70.0.0
six==1.16.0
soupsieve==2.5
spotipy==2.25.1
Pillow==10.4.0
SQLAlchemy==2.0.30
sqlcipher3==0.5.2
stack-data==0.6.3
termcolor==2.3.0
tinycss2==1.3.0
tomlkit==0.12.5
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.0
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
yaspin==3.0.2
